+++ lib/home_page/home.dart
@@ -1,5 +1,5 @@
 import 'package:ckcstudent/home_page/widgets/category_list/category_list.dart';
-import 'package:ckcstudent/home_page/widgets/image_carousel.dart';
+import 'package:ckcstudent/home_page/widgets/image_carousel/image_carousel.dart';
 import 'package:ckcstudent/home_page/widgets/top_banner/top_banner.dart';
 import 'package:ckcstudent/models/app_config.dart';
 import 'package:ckcstudent/widgets/right_appbar.dart';
