--- lib/home_page/widgets/image_carousel.dart
@@ -1,92 +1,0 @@
-import 'package:cached_network_image/cached_network_image.dart';
-import 'package:carousel_slider/carousel_slider.dart';
-import 'package:flutter/material.dart';
-
-class CarouselWithIndicator extends StatefulWidget {
-  final List<String> imgList;
-
-  CarouselWithIndicator({this.imgList});
-
-  @override
-  State<StatefulWidget> createState() {
-    return _CarouselWithIndicatorState();
-  }
-}
-
-class _CarouselWithIndicatorState extends State<CarouselWithIndicator> {
-  int _current = 0;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(children: [
-      CarouselSlider(
-        items: widget.imgList
-            .map((item) => Container(
-                  margin: EdgeInsets.symmetric(horizontal: 5),
-                  child: ClipRRect(
-                    borderRadius: BorderRadius.circular(8),
-                    child: Container(
-                      height: 200,
-                      child: CachedNetworkImage(
-                        fit: BoxFit.fill,
-                        imageUrl: item,
-                        progressIndicatorBuilder:
-                            (context, url, downloadProgress) => Center(
-                          child: CircularProgressIndicator(
-                              value: downloadProgress.progress),
-                        ),
-                        errorWidget: (context, url, error) => Icon(Icons.error),
-                      ),
-                    ),
-                  ),
-                ))
-            .toList(),
-        options: CarouselOptions(
-            autoPlay: true,
-            enlargeCenterPage: true,
-            height: 200,
-            onPageChanged: (index, reason) {
-              setState(() {
-                _current = index;
-              });
-            }),
-      ),
-      Row(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: widget.imgList.map((url) {
-          int index = widget.imgList.indexOf(url);
-          return Container(
-            width: 8.0,
-            height: 8.0,
-            margin: EdgeInsets.symmetric(vertical: 10.0, horizontal: 2.0),
-            decoration: BoxDecoration(
-              shape: BoxShape.circle,
-              color: _current == index
-                  ? Theme.of(context).primaryColor
-                  : Color.fromRGBO(0, 0, 0, 0.1),
-            ),
-          );
-        }).toList(),
-      ),
-    ]);
-  }
-}
-
-class ImageCarousel extends StatelessWidget {
-  final List<String> imgList;
-
-  ImageCarousel({this.imgList});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: EdgeInsets.only(
-        top: 9, // 30 - blurRadius - offset.y
-        bottom: 38,
-      ),
-      child: CarouselWithIndicator(
-        imgList: this.imgList,
-      ),
-    );
-  }
-}
